requirements.txt
Flask==1.0.2
json2html==1.3.0
requests==2.22.0
mysql_connector_repackaged==0.3.1
